requirements.txt
fastapi
slowapi
pydantic
google-generativeai
google-genai
uvicorn
gunicorn
python-dotenv
psycopg2-binary